Update tests to reflect change, update verb tests
Created a fake question api controller preloaded with an
HttpContextAccessor for tests that don't use the context accessor,
updated tests that use a context accessor to create a mock one and
verify its usage. Updated verb tests in AuthApiTests and ApiWriteTests
to loop through a collection of method signatures to test with
HasAttribute.